fix(pki): move mapEnumsForValidation to certificate-constants
certificate-v3-fns is imported by lightweight modules such as
certificate-est-v3-service. Importing mapEnumsForValidation from
certificate-utils pulled isPqcAlgorithm and therefore tsyringe into that
graph, so certificate-est-v3-service.test.ts failed at import time with
"tsyringe requires a reflect polyfill".

mapEnumsForValidation only depends on the two legacy usage mappers, which
already live in certificate-constants (zod only), so it belongs there.
