+++ backend/src/services/certificate-v3/certificate-approval-fns.ts
@@ -42,7 +42,8 @@ import { copyMetadataFromRequestToCertificate } from "@app/services/resource-met
 import {
   CertExtendedKeyUsageType,
   CertKeyUsageType,
-  CertPolicyState
+  CertPolicyState,
+  mapEnumsForValidation
 } from "../certificate-common/certificate-constants";
 import {
   calculateFinalRenewBeforeDays,
@@ -59,7 +60,6 @@ import {
   buildSubjectAlternativeNamesFromTemplate,
   convertExtendedKeyUsageArrayToLegacy,
   convertKeyUsageArrayToLegacy,
-  mapEnumsForValidation,
   normalizeDateForApi
 } from "../certificate-common/certificate-utils";
 import { TCertificateRequestDALFactory } from "../certificate-request/certificate-request-dal";
