+++ backend/src/services/certificate-v3/certificate-v3-fns.ts
@@ -3,7 +3,7 @@ import RE2 from "re2";
 import { BadRequestError } from "@app/lib/errors";
 import { ms } from "@app/lib/ms";
 
-import { mapEnumsForValidation } from "../certificate-common/certificate-utils";
+import { mapEnumsForValidation } from "../certificate-common/certificate-constants";
 import { TCertificateProfileDefaults } from "../certificate-profile/certificate-profile-types";
 
 export const parseTtlToDays = (ttl: string): number => {
